Type conversion change
This change preserves the true type information of Variables and Functions until the last stage of serialization. Previously we converted the type to a WASM binary value_type when generating WAST nodes, but this meant that some stages of the compiler could only see the value_type not the real type. At the current stage this wasn't much of a problem. But later on if we encode pointers as say a i32 it would have allowed pointers to be added to other i32 values...

As part of this change some of the Variablish values have become actual Variable instances for consistency.