Add logical NOT operator
most of the work for this kind of existed already, as the while loop compilation required it for the condition. so this basically just implements the parsing ( which was quite simple ) creates a new compilation function and then calls to a common invert_boolean_expression function which is shared with the while compilation function.