Correct serializer to use PrimativeTypes
originally the serializer had lots of switches that used the string values of the types to decide what instructions to generate. as AtiumType is no longer a string this has had to change. the new atiumtype has a method that is responsible for returning the primative type ( as an enum ) or throwing if that is not possible. the switches now use the enum for switching where appropriate. In addition the value stack now uses primative types instead of atium types, as it's mostly a validation system for the primative types and not responsible for the actual type checking system used in the compiler ( but which the serializer is aware of ).

additionally in those switch statements each now has a default statement that throws a compiler error ( they SHOULD be unreachable if the compiler has worked correctly, hence why it is a compiler error ).