Parser now generates a TypePattern instead of a string for types
This change was made to allow the parser to emit more complex types than just a name. Such as tuples, functions and arrays. Initial support is only for classes and tuples. Unfortunately this change also modifies the compiler a bit as it inserts various types into the IR that the user does not specify ( such as equality expressions resolving to booleans ).

While doing this we also migrate the use of compiler inserted types to use static type objects, instead of requesting new instances from the type parser.