Wrap exported functions
This change facilitates us to wrap calls to an exported function when they come from the host environment. The intention is to use this to manage the temporary heap allocation reset system.

In implementing this the ability to create a function that is not visible to the user, and is treated as unique even when it has a name that collides with another. This will likely be used later for lambdas and methods.