Add lazy created variables to function environment for array acces
Doing the bounds check on an array access requires 3 temporary variables. One for the array pointer, one for the array index and one for the array length. It's quite possible that many array accesses are used in a single function, so we don't want to be creating 3 new variables for each one when we can just reuse them. Hence this adds a method to the environment ( which is unique to each function ) that lazy creates these variables and returns them to the caller, so that we can reuse them.