revert change to pass AST object to WAST nodes
WAST nodes contain Reference objects that indicate a position within the source code. Previously we had created these as early as possible and passed them down into varoious sub-functions to encourage deduplication of these references. However, we ended up with *alot* of additional function parameters because of this. Hence a previous commit changed the code to ditch these parameters in favour of creating the Reference within the WASTNode constructor from a AST object. Many of the compilation functions involved passing the node in anyway so this greatly simplified code. It did however mean that in a number of cases we were passing nodes into functions that only wanted the position data from the node.

This change is a partial reversion of that commit. The compilation structure has changed so a straight reversion is impractical. Also now we are creating the reference object when a function needs it, as opposed to as early as possible. So we don't have the explosion of functions parameters that we did last time.