tidy: change AST export from type alias to reexport
Previously the pratt Node type was being reexported by core as the type alias AST. The name change being mostly down to name collisions with the HTML Node object which TS kept on getting confused by. As the type alias was a simple a = b job it made sense to just reexport here, to simplify differences between type aliases and actual types.