change: modify pratt.Node to be an optional generic
The data value of Node has historically been "unknown" and then cast to the actual type by the consumer. This is less than ideal for the type checker. This change makes the data value generic, with a default of unknown. Allowing us to add a type to the data in certain areas. With adjustment we may be able to get complete type verification of the data field from parser to compiler without needing any type casts.