fix: possible invalid depth assertion in list
list expressions assert the stack depth before exiting to ensure nothing has been left on the stack incorrectly. A list that emits a value is allowed a stack depth of 1, but one that doesn't is allowed a stack depth of 0.

The assertion incorrectly assumed the stack depth when entering the list was 0 ( which it commonly is, but not always ). This changes to compare the stack depth delta instead.