tidy: remove as_[type] in favour of is_[type] with type guards
LangType is a union of various types. Conversion was done using various as_tuple style methods that would return the type as a Tuple if it was one, or as null if it wasn't. While this worked reasonably well it was also being misused as a check in places. This change instead uses type guards, that basically assert that if the method returns true that it is a specific type. Allowing us to cast types using a is_typle style check instead. Which is much simpler.