add: allow struct/tuple casts while in unsafe mode
This adds a new case to the "as" expression allowing for the conversion of raw pointers to and from structs, tuples, arrays and strings. This used to produce an error ( rightly so ) but now it is available in "unsafe" mode.

This is basically a requirement if we have in language memory mangagement techniques. Depening on how the memory system end up being developed this may be removed in favour of inbuilt behaviours ( although these will be harder to implement ).