add: "never" primative type
This type is intended to be used for flow analysis. If an expression is a "never" type it indicates that once it has been evaluted it's execution will stop. The main reason why this is being added is to allow explicit return statements. But it should also be useful for "break" statements in loops, etc.

It is treated as "void" for most purposes, because it also represents the abscense of value. Hence "is_void" will also return true if the type is "never".