fix: invalid expression value for conditionals without "else" clause
The value of conditional expressions defaulted to the value of the "then" clause, then became "void" if the "else" clause had a different type. This mistake meant that conditions without an else clause appeared to have a value, but they could have been void.

It now defaults to "void" and takes on a value if the 2 clauses are present and match.