fix: error loading/storing a string
This was a rather unpleasant error. There was no default case checking for errors in the serialize functions for load and store.

Cases were not added for the str primative type, hence the opcode was silently skipped for these. Creating a confusing error when the module was initialised, as the mem arguments were read as ops.

The serialisation functions need to be reviewed for edge cases like these at some point, because even if all cases are covered at the moment later additions could cause similar bugs.