refactor: unify local vs global variable WASTNodes
Replaces the get/set/tee local/global WASTNodes with get/set/tee variable WASTNodes. So now methods using variables don't have to care if the variables are global or local.

This additionally introduces a new variable WASTNode called var_restore. This wraps an expression with a "save"/"restore" style pattern for a variable, ensuring that the value afterwards is the same as before. This is useful if you expect the expression to modify the variable somehow.