tidy: change various unique IDs from numbers to Symbols
Functions, Variables and Memory all use ID values for lookups within the serialiser which requires indices. Prior to this change counters were used to choose an incrementing integer for the ID. This swaps to using Symbols instead, which simplifies and reduces the chances for bugs.

We could plausibly use the objects themselves instead of Symbols, but that would complicate the object structure somewhat.