change: rework scoping
This is a rather broad and shallow change. It modifies how the context and environment behave regarding scope. All declaration types now have a scope, whereas before only variables were scoped and others were allowed only at the global scope.

This in theory allows for declaring enums/structs/functions within a function scope, or block scope. Although the various other compiler work hasn't been done to allow this.

Additionally this changes how expressions and statements outside of functions behave. Previously only global variables could be created outside of functions, despite the fact they are initialised inside a special start function. This either complicated or broke various code that required temporary variables. Now there is a seperate environment used when outside of functions, which allows for proper scoping of variables. Any top level variables are considered globals and can be accessed by functions. All other are considered locals and can only be accessed within their scope of the start function.

One final change is the introduction of a TypeAlias declaration. This allows a complex type to be assigned to a name. This hasn't been exposed to users yet, but will hopefully be a handy shortcut for implementing templates.