add: option to find_common_type to ignore "never"
`find_common_type` is used effectively for 2 things. First is to evaluate control flow, such as an if expression. Second is to find the common type of multiple values. In the first case a "never" value occurs conditionally, hence the common value is non-inclusive of that "never". In the second case it is non-conditional, and due to the way "never" works it becomes the sole type of the value. This change adds a parameter to the function to reflect these 2 varieties.