add: introduce "break" wast node
Adds a new "break" node, which operates in a similar way to branch nodes except it emits a value ( not directly, but via a surrounding block ). Additionally makes some adjustments to 'block' nodes, to decouple the emitted value of the inner list from the value emitted from the block. This is important; provided the block _always_ breaks/traps it is not required that a value is present on the stack in the inner block.